Rename `Worker` to `ConnectionHandler`
This renames the `Worker` to `ConnectionHandler`. Worker was a
poor name for this piece of logic and comes from assuming the
internals of DatTCP. This renames it to more accurately describe
what it does, handles a new connection. This means parsing a
request, running its route and writing the response. This doesn't
change any functionality, its just a rename.