Split service handler tests into unit and system tests
This updates the testing for service handler to split the existing
tests into unit and server tests. The unit tests no longer use the
test runner to test the service handler. They are strictly unit
tests testing the behavior, but not testing how a service handler
would typically be tested. This is now done in the system tests for
service handlers. In these tests, the test runner is used to run
the same handlers used in the server system tests. This ensures
that service handlers (and their runners) work as expected in
both live and testing scenarios.

Since we switched to using the app server handlers, I was able to
remove the `test/support/service_handlers` file. Both of these
files had the same "kinds" of service handlers. I wanted to have
the same handlers used in a live and testing scenario, so I
decided to re-use the app server handlers and remove the other
handlers.