Add the server ip and port to the process name
This adds the server ip and port to the process name. This was
accidentally removed in the last release of Sanford. Since the ip
and port can be overriden, its nice to include them in the process
name to quickly identify which process is running on an ip/port.
This restores that and also exposes reading a servers configured
ip and configured port instead of its "live" ip and port. The
configured is always based off what the server was configured with
and the "live" changes based on what args were passed to listen.

/cc @kellyredding