server config: update `to_hash` to include `init_procs`
This is a minor cleanup to the server configuration `to_hash`
representation.  This removes `template_source` as it is included
by being a ns option.  This also adds init procs as this is more
proper since we are already adding error procs.  Now the custom
to hash method just adds in non-option atts values.

This also updates a few version dependencies to get on the latest
stuff.

Overall, no behavior changes.  Nothing is currently relying on
init procs being in the hash and template_source is still in it b/c
it is an option.