Fix a bug when inspecting a service handler
This fixes a bug when inspecting a service handler. The previous
implementation called a private method with `self.` which caused
it to throw an error. This didn't fail in the tests because they
were previously making the request, params and logger public.
This changes the tests to not rely on making these public so the
inspect method would fail. This was a particularly annoying
issue when testing handlers because they are inspected by
default.