update server data to include all of the server config data; cleanups
This updates the server data to include *all* of the server config
data.  This includes both the init procs and the router.  All data
should be available in server data as it is passed down via the
connection handler and route to the runner/handler.  This allows
for future features on this data without having to later pipe them
down.

This also follows the convention in Deas of passing all its data
down to the handlers.

No behavior changes b/c nothing currently relies on this data being
there.  This is just making things consistent and not arbitrarily
choosing what data is included.  This also has a bunch of test cleanups.
Attributes previously were not ordered consistently between the server
config and server data.  Also there were some missing attribute
tests (like template source, etc).