runners: switch to getting standard args from a passed in hash
This updates the runners to set their standard args from a given
hash.  This means all runners get their settings similarly and
don't reimplement gathering their settings.  This alson lines up
with how Deas/Qs handle their runners.

This is just a cleanup to make the code easier to understand and
work with.  This also makes it easier to pass down data from the
route to the runners.  No behavior changes.