Change restart cmd to check if `ENV['PWD']` is set
This changes the restart cmd to check if `ENV['PWD']` is set
before it tries to stat it. This covers the rare case where the
env var isn't set, like when using sudo. This piece of code is
heavily influenced by puma and I copied this from them as well.

This also updates the `RestartCmd` and changes its `exec` method to
just `run`. This is simpler and fits more of our standard pattern
of initializing and running classes. This also reworks the tests to
properly test all the behavior with restart cmd. This wasn't done
previously most likely because it is relatively difficult to test
because of the number of stubs and because its mostly edge case
behavior.