Hotfix: Stop defaulting `@service_handler_ns` to `nil` in `Router`
This changes `Router` to no longer manually set its
`@service_handler_ns` to `nil`. This caused problems with
inheriting from the `Router` class and trying to default the
service handler ns. Since all ivars are defaulted to `nil` this
isn't required so this removes it to avoid the issues.

This also fixes a bug where the service handler ns method was
setting a view handler ns ivar instead of a service handler ns
ivar.

/cc @kellyredding