Update `TestRunner` to normalize test params
This updates the `TestRunner` to normalize test params. This
stringifies them and then encodes and decodes them. The goal is
to ensure the params are valid and would work in a real request.

This also updates to work with the latest sanford-protocl. This
involved renaming the `BadRequestError` to `Request::InvalidError`.