Update runner to build its handler after settings its own attrs
This is a small change that moves building the handler after the
runner has set its attributes. This is because the handler takes
the runner when it is initialized. If it wanted to use the
attributes from the runner in its initialize it wouldn't be able
to. By moving building the handler to the end of the runners
initialize the runner is "ready to go" by the time it passes it
to the handler.

This is a small change that came from Qs. Qs handlers needed to
user their runner in their initialize. This is to be consistent
and because its a better default since the handler is passed the
runner. It avoids the scenario where, here is the runner, but you
can't use it yet cause we haven't finished initializing it.

This also updates to the latest version of assert.