Move ip and port env vars usage into server configuration
This moves the ip and port env vars from the process into the
server configuration. The benefit of this is the server data now
always knows the actual configured ip and port. Previously, if the
env vars were set, then the server data could return an incorrect
ip and port than what was actually being used. This also keeps the
same behavior of allowing the env vars to overwrite the servers
configuration.

In addition to the env var changes, the server data was also
updated to have its ip and port once the server starts listening.
This is because there was another scenario where the server data
could return incorrect ip and port information. If the server file
descriptor env var is used then all ip and port configuration is
ignored. So it is possible for the server file descriptor to use a
different ip and port than what was configured which means the
server data would return incorrect data. By writing the ip and port
back to the server data after listening, we know the actual ip and
port being used. The server file descriptor is primarily used for
restarting a sanford process so its not likely that it won't match
what was originally configured but it is possible.