Allow setting dat-tcp worker callbacks
This adds configuration and DSL methods for adding dat-tcp worker
callbacks for start, shutdown, sleep and wakeup. This allows
adding custom callbacks when the previous events happen on a worker
thread.

This will be used to set the thread id of any new workers on our
logger so each worker will log with its own thread id.