Fix using an ip and port env vars
This fixes a bug with using ip and port env vars to configure
the ip and port the server listens on. The ip and port env vars
are intended as a way to overwrite any configuration and tell a
process to have its sanford server listen on an ip and port. The
previous implementation instead used them as a default. This
meant that the default value would be based on the env var but if
the servers configuration was changed it would be overwritten.

This fixes the bug by reverting to the old default ip and port
configuration logic and having the server data check for the ip
and port env var when its built. If it sees they are set, it will
use the env var instead of whats configured.

This also adds a system test to ensure that if we create a server
with the env vars set it uses them when binding to a socket. This
should ensure that if this logic moves the overall behavior still
works.

Finally, this also removes an old process test that used the ip
and port. This was a remnant of when the process would look for
the ip and port env vars and is no longer needed.