move test helpers into the service handler source files
This simplifies the number of files you need to remember to require
and more accurately organizes the test helpers.  This also adds
the missing test helpers tests.

Note: this also adds a custom equality methods for handlers and
tests it.  This should have been done originally but wasn't needed
until now since we are testing the test helpers.