switch to declarative response handling
This updates the handlers to declare their response values.  In the
handler, the return value is no longer used to affect the response.
You declare status (code/msg) and data changes via private helper
methods.  The goal here is to be able to change the response in
the callbacks.

As part of this effort, the runners have been updated to support
this behavior and to make their handling more closely match Deas'.
This involved a number of smaller changes, including:

* the sanford runner now catches halt in its before/after callbacks
* added status/data helpers to change response values
* halt and status both take similar args
* the render method auto sets the data
* the test runner no longer has a response_value reader (use the
  run method's return value instead)