Pass num workers, logger and shutdown timeout to dat-tcp
This updates sanford to pass num workers, logger and shutdown
timeout to dat-tcp. The logger is only passed as part of sanford's
updated debug mode which allows extra logging by dat-tcp. As part
of this a sanford server can now be configured with a number of
workers and a shutdown timeout. This is consistent with qs,
formalizes sanfords debug mode and makes it more flexible.

A server can now be configured with a num workers and a shutdown
timeout. These values will be passed onto dat-tcp and will
determine how many workers are initially spawned and how long
it lets its shutdown logic run. Qs allows configuring these so
this makes sanford consistent with it.

The server data now also has a debug flag that is set by the env
var `SANFORD_DEBUG`. When this is set, the sanford logger will be
passed to dat-tcp. This shouldn't be set in a live environment
because dat-tcp does a lot of extra logging that will slow down
sanford handling connections.

This also removes the connection handler raising an exception if
the `SANFORD_DEBUG` flag is set. This is a development feature that
was used when initially developing sanford logic. It was never
intended to be a feature outside of the sanford repo and test suite.
This removes it so its no longer tied to the `SANFORD_DEBUG` env
var. If we need it again when updating sanford we can figure a way
to add the feature back so that it is only part of the test suite.