Update to use the new dat-tcp
This updates sanford to use the new dat-tcp. dat-tcp's interface
has changed but functionally it works the same as it previously
did. This also brings in much-plugin and uses it for its server
and service handler mixins. This ensures the mixin logic is only
included once.

A server can now be configured with a worker class instead of
worker procs. This is because dat-tcp now requires a worker class
when its initialized. A block is no longer part of dat-tcp's
interface. The logic for processing a work item (a client socket
for sanford) is now part of the worker class that is passed in.
Worker callbacks should now be added to the worker class thats
configured.

As part of this change, sanford now also provides a `Worker` mixin
which as mentioned before, contains the logic for processing a
client socket. This mixin is required for any worker class that is
configured on a server.

This also updates styles and conventions to our latest standards
throughout and updates the system tests to do a better job of
looking for an non-localhost IPv4 address. The previous code could
return an IPv6 address.