Pass registered engine ext to engine as option
This updates template sources to pass the extension that was used
to register an engine to the engine as an option. The goal is to
allow the engine to choose templates that match the extension that
was used when configuring the extension. This is also part of
fixing an issue with the engines not knowing which template file
to pick because they no longer hard-code an extension that the
templates have to use.

The template source now always sets an ext option and passes it
to the engines when they are registered. The ext will always be
set and can't be different from how the engine is registered. It
would be non-sensical to register with one engine ext and pass a
different engine ext as an option. Sanford would either not use
the registered engine to render the template or the template
engine wouldn't find the file. Either case wouldn't produce the
desired result.

Specifically, this is part of fixing an issue with the nm template
engine no longer forcing templates extension to be `nm` (see
redding/nm#16). By not expecting an extension, the nm engine can't
pick the correct file as the template if there are multiple files
with the same name except for their extensions. For example, a
common pattern is to name a handler file and the template the same
only the handler will have `.rb` for its extension and the template
would use `.nm`. In this scenario, it's possible for the nm engine
to try and use the ruby file as its template. Nm is being updated
to use a custom extension (the one sanford will now pass to the
engine) to address this issue (redding/nm#17).

Finally, this also renames "path" variables to "name" in the
template source and engine. The variables are expected to be names
and not full paths. This doesn't change any functionality it just
clarifies the variable names.