Raise argument error if template name matches multiple files
This updates the template source and null template engine logic
to raise an argument error if the template name matches multiple
template files. This also changes how the template source glob
for template files when parsing the ext to determine the engine it
should use. It no longer expects templates to have extensions and
no longer rejects if the templates extension is not a configured
extension. All of this is to try and make Sanfords template
rendering more straightforward and avoid surprises.

By raising an error if multiple template files match, this avoids
any surprises with an unexpected file being rendered. This is
particularly nefarious if a development machines orders files one
way when globbing and a production machine orders them a different
way. In development everything would work as expected but then
fail in production.

By not expecting a `.` after the template name and not rejecting
template files that do not match a configured engine, the
rendering is now more flexible. If a user wants to render a static
file (that doesn't need an engine) they can and the null template
engine will automatically take over for them. The template file
name also doesn't have to match an imposed convention. Finally,
this also sets up Sanford to allow using multiple template engines
in the future if needed.

Finally, this no longer disallows extensions for engine
registration. Previously, we would explicitly not allow rb engines
which would ensure that we didn't pick an rb file as a template.
This is really restrictive and doesn't really provide any
protection. If a user wants to use an template file with an rb
extension they should be allowed to.