templates: cache null template engines on known null extensions
This minor optimization will now cache null template extensions.
The idea is to not repeatedly call the engine cache default block
for known null extensions.

To make this possible, I had to update the logic for determining
if we have a registered engine.  We now store a separate list of
registered extensions. This implementation is nice b/c it decouples
the engine cache from defining whether an engine is registered or not.

Note: this optimization matches Deas' handling of its null template
engines.  See redding/deas PR 210 for details.