remove ns-options dependency
We are moving away from using ns-options as a dependency b/c it
has performance issues when accessing options and on top of that
its API hasn't proven to be especially useful over a struct-like
class.

This removes ns-options and reworks the server, its config and the
server data to support the non-ns-options config implementation.
The config is now a simple struct-like class with a bunch of
accessors for its attrs.  The server class methods now implement
the ns-options-like DSL api for each of their attrs for backwards
compatibility.

In addition, I did a number of smaller changes/cleanups, including
some thing that aren't backwards compatible:

* switched to calling the config "config" (more succinct than prev
  with no loss of clarity).
* you can no longer set config values to procs (for lazy eval'ing).
  I chose to not keep this behavior.
* I dropped the `to_hash` logic and switched to just manually
  building the server data.  I'm just trying to simplify the API
  here and blindly using the config's hash kinda hides what values
  are being set.
* I reworked the order of config/server/server-data settings to
  try and get everything to match.  This is just to be consistent
  and help keep things organized/readable.  I tried to follow the
  order Deas uses as much as possible.
* I did a bunch of test cleanups to not only test the new logic
  but use modern conventions like Factory values and also test
  things more thoroughly.
* updated to the NullTemplateSource to take an optionall root path.
  This matches Deas' implementation and also allows default template
  source to be based on the pwd (like Deas' is).  This is a more
  usable default.
* updated the router to abstract its `validate!` logic to match
  Deas' implementation.
* updated the assert dependency to get access to the latest stuff.

Overall there are some backwards incompatible changes but the basic
behavior isn't changing.  Most of the diff is just reordering things
and reworking to remove ns-options logic where needed.