cleanup server/server-data implementation to better match Deas/Qs
The goal here is that we keep all 3 of there "interface" gems in
sync implementation-wise as much as possible.  These are a bunch
of cleanups that I noticed while removing the ns-options dependency
from Qs.  I used Sanford for reference as it has already removed
its ns-options dependency (see PR 173).

Most of this stuff is relatively minor, but I want to highlight
two not-so-minor changes that are backwards incompatible:

* I chose to remove `init_procs` from the server data.  This is
  consistent with how Deas/Qs don't have this in their corresponding
  server/daemon datas.  The init procs are designed to be run once
  at server init time are run from the server's main Config.  They
  don't need to be passed around in the server data as nothing that
  receives the server data will need the init procs.
* I chose to remove the `dat_tcp_server` reader from the server
  instances.  This, again, matches how Qs does not expose its
  `worker_pool` on its daemon instances.  This is an implementation
  detail that we don't need to expose - plus this keeps the server
  api as simple as possible.

All the other changes are either style-related, consistent attr
order implementations and tests, or are updates to keep consistent
with Qs' implementation.  Most of these *should* have been done in
PR 173 but were missed.