get test suite passing in modern ruby versions
This updates the gem dependencies to bring in the latest versions
(which also now work in modern ruby versions) and updates the test
suite to get things passing in modern ruby versions. This mostly
updating response handling to use response codes (as this is the
only way to build responses in the latest sanford-protocol).

In addition, I had to update the tests' AppServer and its ERB
handling.  Apparently, calling `send(:binding)` on an object no
longer returns a binding in the scope of the object.  So, I had
to hold Ruby's hand and go through the theatrics of setting up
a `get_binding` method on the RenderScope that calls binding
from said method.  Ok Ruby, I'll be super formal about it all.