update the process to label more like Qs does
This does a number of small tweaks to get Sanford setting its
process label like Qs does.  This label is slightly easier to read.
This also adds the ability to override the label with an ENV var.

The process label is now set by the server data, proxied by the
server where the process accesses and uses it.  The "sanford" str
is no longer part of the raw label - it is now added by the process.
This allows you to override the details of the label with the env
var while retaining the `sanford:` label prefix.