update signal handling to use self-pipe; mimic Qs' implementation
This updates Sanford's signal handling to avoid doing "unsafe"
things in trap handlers.  The idea is you want to put as little
logic into the traps as possible, both to make them responsive
and b/c Ruby in 2.0+ starts failing if it detects "unsafe" things
in traps.

Note: Ruby 2 *should* have been complaining about the previous
implementation b/c a mutex sync was happening in the traps. I'm
note sure why it wasn't (maybe it was a few layers deep and that
confused the interpreter?) but this removes that "unsafe" action
and replaces it with a self-pipe implementation.

Now the traps just write signals onto the self pipe while the main
thread polls the pipe for signals to process.  Any processing
happens in the main thread instead of the traps.

This also reworks the process tests to support this new scheme
and makes them more thorough (this is just a mimic of Qs' tests).