Add `running?` and `listening?` methods to `Server`
This adds a `running?` and `listening?` method to the `Server`
which fixes an error that was happening when running a sanford
process. The process expects a `running?` method to check if the
server's thread has crashed for any reason. This passed previous
tests because the server spy used in the tests had a `running?`
method but `Server` didn't. This adds it to the server to fix
this issue.

This also adds `listening?` because its very similar to `running?`
and its consistent to either have all of the methods for checking
the server's state or have none of them.

This was noticed while running the bench script to test if sanford
works in newer versions of ruby. This also includes an updated
bench report for ruby 1.8.7 because of this.