Catch non `StandardError` exceptions in `ConnectionHandler`
This updates the `ConnectionHandler` and `ErrorHandler` to catch
non `StandardError` exceptions. This updates sanford to rescue the
same exceptions that dat-worker-pool workers will rescue. These are
exceptions that we want to run through the configured error procs
even though they don't inherit from `StandardError` because they
won't shutdown the worker thread but they do cause a service to
fail. This includes ruby's `Timeout::Error`, `LoadError` and
`NotImplementedError`.

This updates the `ConnectionHandler` to not only rescue
`StandardError` but instead use dat-worker-pool's standard error
classes. Just like all standard errors have been up to this point,
any of these errors will be passed to the error handler which will
run through the configured error procs.

The `ErrorHandler` also rescues dat-worker-pool's standard error
classes while it is running configured error procs. This is
intended to be consistent with the payload handler; we want to
rescue the same kind of errors that happen while trying to run an
error proc. This ensures we are consistent in our error handling
while processing a service request.