add `partial` method to handlers and template source/engine
This is an alternate to the `render` method that doesn't pass the
service handler to the template as a local.  Use this to render
partial templates that only rely on locals.

This is also to try and line up the template source api with Deas'
template source api.  Libs that provide common rendering behavior
for both Deas and Sanford expect the template source api to be the
same so this makes that happen.

Specifically, this will allow adding common picker items data
rendering logic needed by Romo.  Option data can come both from
a Deas view handler or from a Deas view handler via a Sanford
service call.  In both cases, we want to render the picker items
data in a common way.  This way, no matter whether we are given a
Deas or a Sanford template source, we can call `partial` in our
Romo common logic to render a common partial template for the
picker items data.